Modification to database interface
