--- SmartReceipts/Persistence/Helpers/DistancesHelper.h
@@ -1,14 +1,0 @@
-//
-//  DistancesHelper.h
-//  SmartReceipts
-//
-//  Created by Jaanus Siim on 28/04/15.
-//  Copyright (c) 2015 Will Baumann. All rights reserved.
-//
-
-#import <Foundation/Foundation.h>
-#import "Database.h"
-
-@interface DistancesHelper : Database
-
-@end
